+++ lib/geolix/record/represented_country.ex
@@ -7,6 +7,7 @@ defmodule Geolix.Record.RepresentedCountry do
 
   defstruct [
     :geoname_id,
+    :is_in_european_union,
     :iso_code,
     :name,
     :names,
@@ -14,7 +15,10 @@ defmodule Geolix.Record.RepresentedCountry do
   ]
 
   def from(nil, _), do: nil
-  def from(data, nil), do: struct(__MODULE__, data)
+
+  def from(data, nil) do
+    struct(__MODULE__, Map.put_new(data, :is_in_european_union, false))
+  end
 
   def from(data, locale) do
     result = from(data, nil)